unlpimage/requirements.txt
REQUIREMENTS

For main program:

Python v 3.11.x
https://www.python.org


Python libraries:

PySimpleGUI v 4.60.4
https://github.com/PySimpleGUI/PySimpleGUI

Pillow v 9.5.0
https://python-pillow.org


*-*-*-**-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*

For statistics:

Jupyter core packages:

IPython          : 8.12.0
ipykernel        : 6.19.2
ipywidgets       : 8.0.4
jupyter_client   : 7.1.2
jupyter_core     : 5.3.0
jupyter_server   : 1.13.5
jupyterlab       : 3.3.2
nbclient         : not needed
nbconvert        : 5.5.0
nbformat         : 5.1.3
notebook         : 6.5.4
qtconsole        : 5.3.0
traitlets        : 5.7.1


Python libraries:

numpy v 1.24.3
https://www.numpy.org

matplotlib v 3.7.1
https://matplotlib.org

pandas v 2.0.2
https://pandas.pydata.org

wordcloud v 1.9.2
https://github.com/amueller/word_cloud


